♻️🎨🚚 Refactor logger module.
